Change `amount` type to SMALLINT(5) UNSIGNED

+++ sql-files/upgrades/upgrade_20150305_market.sql
@@ -5,7 +5,7 @@
 CREATE TABLE IF NOT EXISTS `market` (
   `name` varchar(32) NOT NULL DEFAULT '',
   `nameid` SMALLINT(5) UNSIGNED NOT NULL,
-  `price` INT(11) UNSIGNED NOT NULL,
+  `price` SMALLINT(5) UNSIGNED NOT NULL,
   `amount` INT(11) UNSIGNED NOT NULL,
   `flag` TINYINT(2) UNSIGNED NOT NULL DEFAULT '0',
   PRIMARY KEY  (`name`,`nameid`)
